requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
attrs==23.2.0
Brotli==1.1.0
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
comtypes==1.4.4
curl_cffi==0.6.4
customtkinter==5.2.2
darkdetect==0.8.0
Deprecated==1.2.14
freeGPT==1.3.5
frozenlist==1.4.1
g4f==0.3.2.1
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.135.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.1
googleapis-common-protos==1.63.2
grpcio==1.64.1
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.7
mss==9.0.1
multidict==6.0.5
nodriver==0.29rc2
packaging==24.1
pillow==10.3.0
proto-plus==1.24.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.4
pydantic_core==2.18.4
pyparsing==3.1.2
pypiwin32==223
pyttsx3==2.90
pywin32==306
requests==2.32.3
rsa==4.9
tqdm==4.66.4
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
